+++ config/splat.dra.yaml
@@ -55,7 +55,8 @@ segments:
       - [0x418E0, rodata]
       - [0x41990, rodata]
       - [0x419B0, rodata]
-      - [0x419D0, rodata]
+      - [0x419D0, .rodata, 75DA0]
+      - [0x419E8, rodata]
       - [0x41A00, rodata]
       - [0x41F30, rodata]
       - [0x41F70, rodata]
@@ -66,5 +67,7 @@ segments:
       - [0x4768C, c, 4768C]
       - [0x5D7C0, c, 5D7C0]
       - [0x5D874, c, 5D874]
+      - [0x75DA0, c, 75DA0]
+      - [0x75F54, c, 75F54]
       - [0x962AC, data]
   - [0x119870]
